Use wp values for user_id
Return the user_id from the db instead of the user input.
Otherwise case-insensitive collations enable the user to login
with lowercase and uppercase and then multiple users are created in
the SPs where the info is used. Settle on the "one true" representation
from the wordpress db to avoid.